Fix test code's handling of new stats thread.
- CHECKPOINT. UNTESTED merge of this commit from s3hack branch.
- Update terminateTestRun() and waitForThreads() in TestCommunityView
to correctly handle existence of new stats thread.
- Create utils.py and move several utility functions from
communityview.py to utils.py.
- Add thread re-naming functions to utils.py and use for stats thread.
- Improve the way localsettings vars are replaced by testsettings vars
in TestCommunityView.py and testStats.py.
- Misc bug fixes in test code.
- Misc comment fixes.

+++ CommunityView/src/test/TestCommunityView.py
@@ -21,8 +21,15 @@
 ################################################################################
 
 import unittest
-import communityview
 import testsettings
+import localsettings
+
+# Set up the testing values for the global config vars
+#
+localsettings.root                  = testsettings.root
+localsettings.cameras               = testsettings.cameras
+
+import communityview
 import time
 import threading
 import logging
@@ -31,6 +38,8 @@ import shutil
 import inspect
 import datetime
 import platform
+import stats
+from utils import is_thread_prefix
 
 moduleUnderTest = communityview
 
@@ -96,7 +105,8 @@ def deleteTestFiles():
     top-level directory of the CommunityView website (and is also used as the
     top directory for incoming images.
     """
-    shutil.rmtree(moduleUnderTest.root, False, None)
+    if os.path.isdir(moduleUnderTest.root):
+        shutil.rmtree(moduleUnderTest.root, False, None)
     os.mkdir(moduleUnderTest.root)
 
 
@@ -154,10 +164,11 @@ def file_has_data(path):
 def validateWebsite(image_tree):
     success = True
     root = moduleUnderTest.root
+
+    assert file_has_data(os.path.join(incroot, "index.html"))
     
-    assert file_has_data(os.path.join(root, "index.html"))
-    rootdirlist = os.listdir(root)
-    if len(rootdirlist) > len(image_tree)+1:
+    incrootdirlist = os.listdir(incroot)
+    if len(incrootdirlist) > len(inc_image_tree)+2: # +2: index.html, stats
         success = False
         logging.error("Extraneous file(s) in %s: %s" % (root, rootdirlist))
         
@@ -226,6 +237,8 @@ def validateWebsite(image_tree):
 class TestSurveilleance(unittest.TestCase):
 
     origThreadList = threading.enumerate()
+    stats_thread = None
+    stats_run = threading.Event()
 
     def setUp(self):
 
@@ -236,11 +249,6 @@ class TestSurveilleance(unittest.TestCase):
                 pass
         moduleUnderTest.set_up_logging()
         
-        # Set up the testing values for the communityview global vars
-        #
-        moduleUnderTest.cameras = testsettings.cameras
-        moduleUnderTest.root = testsettings.root
-       
         # override the datetime.date().today method
         datetime.date = ForceDate
            
@@ -390,29 +398,41 @@ class TestSurveilleance(unittest.TestCase):
         if threading.currentThread().name == "MainThread":
             self.waitForThreads()   # wait for communityview to complete current tasks
             # if we've had a pass through communityview's main loop without
-            # finding any work to do, set the terminate flag and return
+            # finding any work to do, set the main loop terminate flag and the
+            # stats loop terminate flag, then release the stats thread, and wait
+            # for it to finish
             if moduleUnderTest.images_to_process == False \
                     and moduleUnderTest.files_to_purge == False:
                 moduleUnderTest.terminate_main_loop = True
+                stats.terminate_stats_loop = True
+                self.stats_run.set()    # release the blocked stats thread
+                assert self.stats_thread, "Don't have stats thread to wait on."
+                self.stats_thread.join()     
+        # if this is the stats thread calling sleep(), block until the main loop
+        # is done with its processing, then let stats thread run to write the
+        # stats file(s)
+        elif is_thread_prefix(threading.current_thread(), "Stats"):
+            self.stats_thread = threading.current_thread()
+            self.stats_run.wait()
         else:
             # the only other sleep call is in processtoday().
             # If processtoday() is trying to sleep, it thinks it's done with
-            # its work, so force it to return so that its thread will die
+            # its work, so force it to return so that its thread will die.
             moduleUnderTest.terminate_processtoday_loop = True
 
     def waitForThreads(self):
+        """Wait for all threads to die that are not either threads
+        that were running when the test was started, or the stats thread."""
         wait = True
         while wait:
             wait = False
             for thread in threading.enumerate():
-                if self.origThreadList.count(thread) == 0:
+                if self.origThreadList.count(thread) == 0 \
+                        and not is_thread_prefix(thread, "Stats"):
                     logging.info("waitForThreads: waiting for "+thread.name)
                     wait = True
                     thread.join()
         logging.info("waitForThreads: done waiting for all threads")
-                        
-
-
 
 if __name__ == "__main__":
     #import sys;sys.argv = ['', 'Test.testName']
